Trends against GMST
Added several new features. The main new feature is that trends can now be computed as change per degree change in GMST, as opposed to trend per year (or decade). Also fixed existing problems where certain months were not grouped properly. For example, DJF would not be grouped properly, as it would rather be JFD as opposed to DJF. That issue is solved now.