Added option to regrid data
This patch is quite large as it took a lot of rewriting of existing codes to allow for this change regridding of datasets. I also added two new scripts (TemporalTrendsAllinOne and SpatialTrendsAllinOne). These should be largely ignored. These are more sloppy scripts used purely to make subplots of existing results. The SpatialTrends and TemporalTrends scripts are the main focus and are used to obtain results of individual scenarios. Whereas these 'AllinOne' scripts are used to obtain several results at once and plot them in one figure. This will likely later be changed.